+++ src/main/java/kitchenpos/tablegroup/application/TableGroupValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.tablegroup.application;
 
 import kitchenpos.common.exception.MinimumOrderTableNumberException;
 import kitchenpos.common.exception.NotEmptyOrderTableStatusException;
@@ -7,6 +7,7 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.dto.TableGroupRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
