+++ src/test/java/kitchenpos/ordertable/application/OrderTableValidatorTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.ordertable.application;
 
 import kitchenpos.common.exception.NotEmptyOrderTableStatusException;
 import kitchenpos.common.exception.OrderStatusNotProcessingException;
@@ -8,6 +8,7 @@ import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.product.domain.Product;
 import kitchenpos.tablegroup.application.TableGroupService;
